Rename TruevalAgentBase -> BaseTruevalAgent
