+++ pdr_backend/trueval/test/conftest.py
@@ -34,7 +34,7 @@ def mock_ppss(tmpdir):
 @pytest.fixture()
 def predictoor_contract_mock():
     with patch(
-        "pdr_backend.trueval.trueval_agent_base.PredictoorContract",
+        "pdr_backend.trueval.base_trueval_agent.PredictoorContract",
         return_value=mock_contract(),
     ) as mock_predictoor_contract_mock:
         yield mock_predictoor_contract_mock
