+++ pdr_backend/trueval/test/test_trueval.py
@@ -2,7 +2,7 @@ from enforce_typing import enforce_types
 
 import pytest
 
-from pdr_backend.trueval.trueval_agent_base import get_trueval
+from pdr_backend.trueval.base_trueval_agent import get_trueval
 from pdr_backend.models.feed import mock_feed
 
 
@@ -14,7 +14,7 @@ def test_get_trueval_success(monkeypatch):
             return [[0, 0, 0, 0, 100], [300000, 0, 0, 0, 200]]
         raise ValueError(f"Invalid timestamp: since={since}")
 
-    path = "pdr_backend.trueval.trueval_agent_base"
+    path = "pdr_backend.trueval.base_trueval_agent"
     monkeypatch.setattr(f"{path}.safe_fetch_ohlcv", mock_fetch_ohlcv)
 
     feed = mock_feed("5m", "kraken", "ETH/USDT")
@@ -30,7 +30,7 @@ def test_get_trueval_fail(monkeypatch):
     def mock_fetch_ohlcv_fail(*args, **kwargs):  # pylint: disable=unused-argument
         return [[0, 0, 0, 0, 0], [300000, 0, 0, 0, 200]]
 
-    path = "pdr_backend.trueval.trueval_agent_base"
+    path = "pdr_backend.trueval.base_trueval_agent"
     monkeypatch.setattr(f"{path}.safe_fetch_ohlcv", mock_fetch_ohlcv_fail)
 
     feed = mock_feed("5m", "kraken", "eth-usdt")
