+++ pdr_backend/trueval/test/test_trueval_agent.py
@@ -4,7 +4,7 @@ from enforce_typing import enforce_types
 import pytest
 
 from pdr_backend.ppss.web3_pp import Web3PP
-from pdr_backend.trueval.trueval_agent_base import get_trueval
+from pdr_backend.trueval.base_trueval_agent import get_trueval
 from pdr_backend.trueval.trueval_agent_single import TruevalAgentSingle
