+++ pdr_backend/trueval/test/test_trueval_agent_batch.py
@@ -3,7 +3,7 @@ from unittest.mock import patch
 
 import pytest
 
-from pdr_backend.trueval.trueval_agent_base import get_trueval
+from pdr_backend.trueval.base_trueval_agent import get_trueval
 from pdr_backend.trueval.trueval_agent_batch import TruevalAgentBatch, TruevalSlot
 from pdr_backend.util.constants import ZERO_ADDRESS
